refactor(restapi): move user related in new subpackage
