+++ state.go
@@ -11,6 +11,7 @@ import (
 	"github.com/nyttikord/gokord/guild"
 	"github.com/nyttikord/gokord/user"
 	"github.com/nyttikord/gokord/user/status"
+	"github.com/nyttikord/gokord/user/userrest"
 )
 
 // ErrNilState is returned when the state is nil.
@@ -1231,7 +1232,7 @@ func (s *State) UserChannelPermissions(userID, channelID string) (apermissions i
 		return
 	}
 
-	return memberPermissions(guild, channel, userID, member.Roles), nil
+	return userrest.MemberPermissions(guild, channel, userID, member.Roles), nil
 }
 
 // MessagePermissions returns the permissions of the author of the message
@@ -1255,7 +1256,7 @@ func (s *State) MessagePermissions(message *channel.Message) (apermissions int64
 		return
 	}
 
-	return memberPermissions(guild, channel, message.Author.ID, message.Member.Roles), nil
+	return userrest.MemberPermissions(guild, channel, message.Author.ID, message.Member.Roles), nil
 }
 
 // UserColor returns the color of a user in a channel.
